# Conflicts:
#	.phpstan.dist.baseline.neon
#	composer.lock
#	lib/Varien/Cache/Backend/Database.php
#	tests/unit/Mage/Catalog/Model/UrlTest.php
#	tests/unit/Mage/Sitemap/Model/SitemapTest.php

+++ app/code/core/Mage/Eav/Model/Entity/Abstract.php
@@ -1,23 +1,15 @@
 <?php
 
 /**
- * OpenMage
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is bundled with this package in the file LICENSE.txt.
- * It is also available at https://opensource.org/license/osl-3-0-php
- *
- * @category   Mage
+ * @copyright  For copyright and license information, read the COPYING.txt file.
+ * @link       /COPYING.txt
+ * @license    Open Software License (OSL 3.0)
  * @package    Mage_Eav
- * @copyright  Copyright (c) 2006-2020 Magento, Inc. (https://www.magento.com)
- * @copyright  Copyright (c) 2017-2024 The OpenMage Contributors (https://www.openmage.org)
- * @license    https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
 /**
  * Entity/Attribute/Model - entity abstract
  *
- * @category   Mage
  * @package    Mage_Eav
  */
 abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_Abstract implements Mage_Eav_Model_Entity_Interface
@@ -395,7 +387,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
             $attributeId = $attribute->getAttributeId();
         }
 
-        if (!$attribute->getAttributeCode()) {
+        if (isset($attributeCode) && !$attribute->getAttributeCode()) {
             $attribute->setAttributeCode($attributeCode);
         }
         if (!$attribute->getAttributeModel()) {
@@ -594,6 +586,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
     public function walkAttributes($partMethod, array $args = [])
     {
         $methodArr = explode('/', $partMethod);
+        $part = '';
         switch (count($methodArr)) {
             case 1:
                 $part   = 'attribute';
@@ -629,7 +622,7 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
                     break;
             }
 
-            if (!$this->_isCallableAttributeInstance($instance, $method, $args)) {
+            if (!isset($instance, $method) || !$this->_isCallableAttributeInstance($instance, $method, $args)) {
                 continue;
             }
 
@@ -1578,18 +1571,20 @@ abstract class Mage_Eav_Model_Entity_Abstract extends Mage_Core_Model_Resource_A
 
         $this->_beforeDelete($object);
 
-        try {
-            $where = [
-                $this->getEntityIdField() . '=?' => $id,
-            ];
-            $this->_getWriteAdapter()->delete($this->getEntityTable(), $where);
-            $this->loadAllAttributes($object);
-            foreach (array_keys($this->getAttributesByTable()) as $table) {
-                // phpcs:ignore Ecg.Performance.Loop.ModelLSD
-                $this->_getWriteAdapter()->delete($table, $where);
+        if (isset($id)) {
+            try {
+                $where = [
+                    $this->getEntityIdField() . '=?' => $id,
+                ];
+                $this->_getWriteAdapter()->delete($this->getEntityTable(), $where);
+                $this->loadAllAttributes($object);
+                foreach (array_keys($this->getAttributesByTable()) as $table) {
+                    // phpcs:ignore Ecg.Performance.Loop.ModelLSD
+                    $this->_getWriteAdapter()->delete($table, $where);
+                }
+            } catch (Exception $e) {
+                throw $e;
             }
-        } catch (Exception $e) {
-            throw $e;
         }
 
         $this->_afterDelete($object);
